Added property OctoPackSkipIfNoChanges
Rewrote OctoPack.targets to be able to create a new Nuget package only if changes have been made to the files that go in to the NuGet package. Just like msbuild will skip projects that have already been built, the purpose of this change is to speed up the build process by not creating a NuGet package with identical files as the previous NuGet package when the property /p:OctoPackSkipIfNoChanges=true is used.